implemented update queries, various code fixes

+++ src/org/uem/dam/gestor_farmacia/model/Medicine.java
@@ -12,7 +12,7 @@ public record Medicine(
 		return new String[] {
 				Integer.toString(article.articleId()),
 				Integer.toString(medId),
-				Integer.toString(mass),
+				Double.toString(mass),
 				unit,
 				Boolean.toString(requiresPresc) };
 	}
